requirements.txt
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
python-speech-features==0.6
tensorflow==2.12.0
transformers==4.30.0
datasets==2.16.1
pyaudio==0.2.14
sounddevice==0.4.6
speechrecognition==3.10.0
pyttsx3==2.90
scikit-learn==1.4.0
pandas==2.1.4
tqdm==4.66.1
pickle-mixin==1.0.2
numpy==1.23.5
torch==2.0.1
scipy==1.11.4